util: move indenter and ansistripper to terminal package
