+++ pkg/pb/interface.go
@@ -29,7 +29,7 @@ func (i *Interface) Device() *wg.Device {
 }
 
 func (i *Interface) Dump(wr io.Writer, verbosity int) error {
-	wri := util.NewIndenter(wr, "  ")
+	wri := t.NewIndenter(wr, "  ")
 
 	if _, err := fmt.Fprintf(wr, t.Color("interface", t.Bold, t.FgGreen)+": "+t.Color("%s", t.FgGreen)+"\n", i.Name); err != nil {
 		return err
