+++ pkg/util/human_test.go
@@ -6,12 +6,13 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"riasc.eu/wice/pkg/util"
+	t "riasc.eu/wice/pkg/util/terminal"
 )
 
 var _ = Context("human", func() {
 	Context("duration", func() {
 		DescribeTable("test", func(dur time.Duration, output string) {
-			Expect(util.StripANSI(util.PrettyDuration(dur))).To(Equal(output))
+			Expect(t.StripANSI(util.PrettyDuration(dur))).To(Equal(output))
 		},
 			Entry("plural", 5*time.Hour+15*time.Minute+2*time.Second, "5 hours, 15 minutes, 2 seconds"),
 			Entry("singular", time.Hour+time.Minute+time.Second, "1 hour, 1 minute, 1 second"),
@@ -22,14 +23,14 @@ var _ = Context("human", func() {
 	Specify("ago", func() {
 		now := time.Now()
 
-		Expect(util.StripANSI(util.Ago(now))).To(Equal("Now"))
-		Expect(util.StripANSI(util.Ago(now.Add(-time.Hour)))).To(Equal("1 hour ago"))
-		Expect(util.StripANSI(util.Ago(now.Add(-time.Hour - 10*time.Minute)))).To(Equal("1 hour, 10 minutes ago"))
+		Expect(t.StripANSI(util.Ago(now))).To(Equal("Now"))
+		Expect(t.StripANSI(util.Ago(now.Add(-time.Hour)))).To(Equal("1 hour ago"))
+		Expect(t.StripANSI(util.Ago(now.Add(-time.Hour - 10*time.Minute)))).To(Equal("1 hour, 10 minutes ago"))
 	})
 
 	Context("bytes", func() {
 		DescribeTable("test", func(bytes int, output string) {
-			Expect(util.StripANSI(util.PrettyBytes(int64(bytes)))).To(Equal(output))
+			Expect(t.StripANSI(util.PrettyBytes(int64(bytes)))).To(Equal(output))
 		},
 			Entry("without SI suffix", 500, "500 B"),
 			Entry("without SI suffix", 1536, "1.50 KiB"),
@@ -39,7 +40,7 @@ var _ = Context("human", func() {
 
 	Context("every", func() {
 		DescribeTable("test", func(dur time.Duration, output string) {
-			Expect(util.StripANSI(util.Every(dur))).To(Equal(output))
+			Expect(t.StripANSI(util.Every(dur))).To(Equal(output))
 		},
 			Entry("plural", 5*time.Hour, "every 5 hours"),
 			Entry("singular", time.Hour, "every 1 hour"),
